cleanup and fixed single precision bug

+++ psi4/src/export_wavefunction.cc
@@ -77,7 +77,7 @@ void export_wavefunction(py::module& m) {
         .def(py::init<std::shared_ptr<Molecule>, std::shared_ptr<BasisSet>, std::map<std::string, std::shared_ptr<Matrix>>,
                       std::map<std::string, std::shared_ptr<Vector>>, std::map<std::string, Dimension>,
                       std::map<std::string, int>, std::map<std::string, std::string>, std::map<std::string, bool>,
-                      std::map<std::string, float>>())
+                      std::map<std::string, double>>())
         .def("reference_wavefunction", &Wavefunction::reference_wavefunction, "Returns the reference wavefunction.")
         .def("set_reference_wavefunction", &Wavefunction::set_reference_wavefunction, "docstring")
         .def("shallow_copy", take_sharedwfn(&Wavefunction::shallow_copy), "Copies the pointers to the internal data.")
